# [WIP] Optimize GetLabels and GetLabelsValues