Using k-merge to merge responses from different ingesters:

This is still a wip: DO NOT MERGE.

Preliminary bench results:

```
goos: linux
goarch: amd64
pkg: github.com/cortexproject/cortex/pkg/distributor
cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
                                           │   /tmp/old    │              /tmp/new               │
                                           │    sec/op     │    sec/op     vs base               │
Distributor_LabelsCommons/10_ingesters-32     61.14m ± ∞ ¹   17.79m ± ∞ ¹  -70.91% (p=0.008 n=5)
Distributor_LabelsCommons/50_ingesters-32     435.5m ± ∞ ¹   101.0m ± ∞ ¹  -76.81% (p=0.008 n=5)
Distributor_LabelsCommons/100_ingesters-32   1012.2m ± ∞ ¹   198.6m ± ∞ ¹  -80.38% (p=0.008 n=5)
geomean                                       299.8m         70.92m	   -76.35%
¹ need >= 6 samples for confidence interval at level 0.95

                                           │   /tmp/old    │               /tmp/new               │
                                           │     B/op	   │     B/op       vs base               │
Distributor_LabelsCommons/10_ingesters-32    6.871Mi ± ∞ ¹   7.269Mi ± ∞ ¹   +5.79% (p=0.008 n=5)
Distributor_LabelsCommons/50_ingesters-32    47.97Mi ± ∞ ¹   38.23Mi ± ∞ ¹  -20.31% (p=0.008 n=5)
Distributor_LabelsCommons/100_ingesters-32   95.80Mi ± ∞ ¹   75.70Mi ± ∞ ¹  -20.98% (p=0.008 n=5)
geomean                                      31.61Mi         27.61Mi        -12.66%
¹ need >= 6 samples for confidence interval at level 0.95

                                           │   /tmp/old    │              /tmp/new              │
                                           │   allocs/op   │  allocs/op   vs base               │
Distributor_LabelsCommons/10_ingesters-32    3926.00 ± ∞ ¹   57.00 ± ∞ ¹  -98.55% (p=0.008 n=5)
Distributor_LabelsCommons/50_ingesters-32    19077.0 ± ∞ ¹   198.0 ± ∞ ¹  -98.96% (p=0.008 n=5)
Distributor_LabelsCommons/100_ingesters-32   38438.0 ± ∞ ¹   391.0 ± ∞ ¹  -98.98% (p=0.008 n=5)
geomean                                       14.23k         164.0        -98.85%
```


**What this PR does**:

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`